+++ src/acom_music_box/__init__.py
@@ -4,7 +4,7 @@ This is the music_box package.
 This package contains modules for handling various aspects of a music box,
 including species, products, reactants, reactions, and more.
 """
-__version__ = "2.3.1"
+__version__ = "2.3.2"
 
 from .utils import convert_time, convert_pressure, convert_temperature, convert_concentration
 from .species import Species
