+++ src/acom_music_box/tools/waccmToMusicBox.py
@@ -6,16 +6,17 @@
 # Author: Carl Drews
 # Copyright 2024 by Atomospheric Chemistry Observations & Modeling (UCAR/ACOM)
 
-#import os
+# import os
 import argparse
 import datetime
 import xarray
 import json
 import sys
+import os
+import shutil
 
 import logging
 logger = logging.getLogger(__name__)
-
 
 
 # configure argparse for key-value pairs
@@ -30,6 +31,8 @@ class KeyValueAction(argparse.Action):
 # argPairs = list of arguments, probably from sys.argv[1:]
 #       named arguments are formatted like this=3.14159
 # return dictionary of keywords and values
+
+
 def getArgsDictionary(argPairs):
     parser = argparse.ArgumentParser(
         description='Process some key=value pairs.')
@@ -45,47 +48,43 @@ def getArgsDictionary(argPairs):
     return (argDict)
 
 
-
 # Convert safely from string to integer (alphas convert to 0).
 def safeInt(intString):
+    intValue = 0
+    try:
+        intValue = int(intString)
+    except ValueError as error:
         intValue = 0
-        try:
-                intValue = int(intString)
-        except ValueError as error:
-                intValue = 0
 
-        return intValue
-
+    return intValue
 
 
 # Convert string to number, or 0.0 if not numeric.
 # numString = string that probably can be converted
 def safeFloat(numString):
-	result = -1.0
-	try:
-		result = float(numString)
-	except ValueError:
-		result = 0.0
+    result = -1.0
+    try:
+        result = float(numString)
+    except ValueError:
+        result = 0.0
 
-	return result
-
+    return result
 
 
 # Build and return dictionary of WACCM variable names
 # and their MusicBox equivalents.
 def getMusicaDictionary():
-   varMap = {
-      "T": "temperature",
-      "PS": "pressure",
-      "N2O": "N2O",
-      "H2O2": "H2O2",
-      "O3": "O3",
-      "NH3": "NH3",
-      "CH4": "CH4"
-   }
+    varMap = {
+        "T": "temperature",
+        "PS": "pressure",
+        "N2O": "N2O",
+        "H2O2": "H2O2",
+        "O3": "O3",
+        "NH3": "NH3",
+        "CH4": "CH4"
+    }
 
-   return(dict(sorted(varMap.items())))
-
+    return (dict(sorted(varMap.items())))
 
 
 # Read array values at a single lat-lon-time point.
@@ -95,51 +94,50 @@ def getMusicaDictionary():
 # modelDir = directory containing model output
 # return dictionary of MUSICA variable names, values, and units
 def readWACCM(waccmMusicaDict, latitude, longitude,
-  when, modelDir):
+              when, modelDir):
 
-  # create the filename
-  waccmFilename = ("f.e22.beta02.FWSD.f09_f09_mg17.cesm2_2_beta02.forecast.001.cam.h3.{:4d}-{:02d}-{:02}-00000.nc"
-    .format(when.year, when.month, when.day))
-  logger.info("WACCM file = {}".format(waccmFilename))
+    # create the filename
+    waccmFilename = ("f.e22.beta02.FWSD.f09_f09_mg17.cesm2_2_beta02.forecast.001.cam.h3.{:4d}-{:02d}-{:02}-00000.nc"
+                     .format(when.year, when.month, when.day))
+    logger.info("WACCM file = {}".format(waccmFilename))
 
-  # open dataset for reading
-  waccmDataSet = xarray.open_dataset("{}/{}".format(modelDir, waccmFilename))
-  if (True):
-    # diagnostic to look at dataset structure
-    logger.info("WACCM dataset = {}".format(waccmDataSet))
-
-  # retrieve all vars at a single point
-  whenStr = when.strftime("%Y-%m-%d %H:%M:%S")
-  logger.info("whenStr = {}".format(whenStr))
-  singlePoint = waccmDataSet.sel(lon=longitude, lat=latitude, lev=1000.0,
-    time=whenStr, method="nearest")
-  if (True):
-    # diagnostic to look at single point structure
-    logger.info("WACCM singlePoint = {}".format(singlePoint))
-
-  # loop through vars and build another dictionary
-  musicaDict = {}
-  for waccmKey, musicaName in waccmMusicaDict.items():
-    logger.info("WACCM Chemical = {}".format(waccmKey))
-    if not waccmKey in singlePoint:
-      logger.warning("Requested variable {} not found in WACCM model output."
-        .format(waccmKey))
-      musicaTuple = (waccmKey, None, None)
-      musicaDict[musicaName] = musicaTuple
-      continue
-
-    chemSinglePoint = singlePoint[waccmKey]
+    # open dataset for reading
+    waccmDataSet = xarray.open_dataset("{}/{}".format(modelDir, waccmFilename))
     if (True):
-      logger.info("{} = object {}".format(waccmKey, chemSinglePoint))
-      logger.info("{} = value {} {}".format(waccmKey, chemSinglePoint.values, chemSinglePoint.units))
-    musicaTuple = (waccmKey, float(chemSinglePoint.values.mean()), chemSinglePoint.units)   # from 0-dim array
-    musicaDict[musicaName] = musicaTuple
+        # diagnostic to look at dataset structure
+        logger.info("WACCM dataset = {}".format(waccmDataSet))
 
-  # close the NetCDF file
-  waccmDataSet.close()
+    # retrieve all vars at a single point
+    whenStr = when.strftime("%Y-%m-%d %H:%M:%S")
+    logger.info("whenStr = {}".format(whenStr))
+    singlePoint = waccmDataSet.sel(lon=longitude, lat=latitude, lev=1000.0,
+                                   time=whenStr, method="nearest")
+    if (True):
+        # diagnostic to look at single point structure
+        logger.info("WACCM singlePoint = {}".format(singlePoint))
 
-  return(musicaDict)
+    # loop through vars and build another dictionary
+    musicaDict = {}
+    for waccmKey, musicaName in waccmMusicaDict.items():
+        logger.info("WACCM Chemical = {}".format(waccmKey))
+        if waccmKey not in singlePoint:
+            logger.warning("Requested variable {} not found in WACCM model output."
+                           .format(waccmKey))
+            musicaTuple = (waccmKey, None, None)
+            musicaDict[musicaName] = musicaTuple
+            continue
 
+        chemSinglePoint = singlePoint[waccmKey]
+        if (True):
+            logger.info("{} = object {}".format(waccmKey, chemSinglePoint))
+            logger.info("{} = value {} {}".format(waccmKey, chemSinglePoint.values, chemSinglePoint.units))
+        musicaTuple = (waccmKey, float(chemSinglePoint.values.mean()), chemSinglePoint.units)   # from 0-dim array
+        musicaDict[musicaName] = musicaTuple
+
+    # close the NetCDF file
+    waccmDataSet.close()
+
+    return (musicaDict)
 
 
 # set up indexes for the tuple
@@ -150,68 +148,138 @@ unitIndex = 2
 # Perform any numeric conversion needed.
 # varDict = originally read from WACCM, tuples are (musicaName, value, units)
 # return varDict with values modified
+
+
 def convertWaccm(varDict):
-  # convert Ammonia to milli-moles
-  nh3Tuple = varDict["NH3"]
-  varDict["NH3"] = (nh3Tuple[0], nh3Tuple[valueIndex] * 1000, "milli" + nh3Tuple[unitIndex])
 
-  return(varDict)
+    moleToM3 = 1000 / 22.4
 
+    for key, vTuple in varDict.items():
+        # convert moles / mole to moles / cubic meter
+        units = vTuple[unitIndex]
+        if (units == "mol/mol"):
+            varDict[key] = (vTuple[0], vTuple[valueIndex] * moleToM3, "mol m-3")
+
+    return (varDict)
 
 
 # Write CSV file suitable for initial_conditions.csv in MusicBox.
 # initValues = dictionary of Musica varnames and (WACCM name, value, units)
 def writeInitCSV(initValues, filename):
-  fp = open(filename, "w")
+    fp = open(filename, "w")
 
-  # write the column titles
-  firstColumn = True
-  for key, value in initValues.items():
-    if (firstColumn):
-      firstColumn = False
-    else:
-      fp.write(",")
+    # write the column titles
+    firstColumn = True
+    for key, value in initValues.items():
+        if (firstColumn):
+            firstColumn = False
+        else:
+            fp.write(",")
 
-    fp.write(key)
-  fp.write("\n")
+        fp.write(key)
+    fp.write("\n")
 
-  # write the variable values
-  firstColumn = True
-  for key, value in initValues.items():
-    if (firstColumn):
-      firstColumn = False
-    else:
-      fp.write(",")
+    # write the variable values
+    firstColumn = True
+    for key, value in initValues.items():
+        if (firstColumn):
+            firstColumn = False
+        else:
+            fp.write(",")
 
-    fp.write("{}".format(value[valueIndex]))
-  fp.write("\n")
+        fp.write("{}".format(value[valueIndex]))
+    fp.write("\n")
 
-  fp.close()
-  return
-
+    fp.close()
+    return
 
 
 # Write JSON fragment suitable for my_config.json in MusicBox.
 # initValues = dictionary of Musica varnames and (WACCM name, value, units)
 def writeInitJSON(initValues, filename):
 
-  # set up dictionary of vars and initial values
-  dictName = "chemical species"
-  initConfig = {dictName: {} }
+    # set up dictionary of vars and initial values
+    dictName = "chemical species"
+    initConfig = {dictName: {}}
 
-  for key, value in initValues.items():
-    initConfig[dictName][key] = {
-      "initial value [{}]".format(value[unitIndex]): value[valueIndex]}
+    for key, value in initValues.items():
+        initConfig[dictName][key] = {
+            "initial value [{}]".format(value[unitIndex]): value[valueIndex]}
 
-  # write JSON content to the file
-  fpJson = open(filename, "w")
+    # write JSON content to the file
+    fpJson = open(filename, "w")
 
-  json.dump(initConfig, fpJson, indent=2)
-  fpJson.close()
+    json.dump(initConfig, fpJson, indent=2)
+    fpJson.close()
 
-  fpJson.close()
-  return
+    fpJson.close()
+    return
 
+
+# Reproduce the MusicBox configuration with new initial values.
+# initValues = dictionary of Musica varnames and (WACCM name, value, units)
+# templateDir = directory containing configuration files and camp_data
+# destDir = the template will be created in this directory
+def insertIntoTemplate(initValues, templateDir, destDir):
+
+    # copy the template directory to working area
+    destZip = os.path.basename(os.path.normpath(templateDir))
+    destPath = os.path.join(destDir, destZip)
+    logger.info("Create new configuration in = {}".format(destPath))
+
+    # remove directory if it already exists
+    if os.path.exists(destPath):
+        shutil.rmtree(destPath)
+
+    # copy the template directory
+    shutil.copytree(templateDir, destPath)
+
+    # find the standard configuration file and parse it
+    myConfigFile = os.path.join(destPath, "my_config.json")
+    with open(myConfigFile) as jsonFile:
+        myConfig = json.load(jsonFile)
+
+    # locate the section for chemical concentrations
+    chemSpeciesTag = "chemical species"
+    chemSpecies = myConfig[chemSpeciesTag]
+    logger.info("Replace chemSpecies = {}".format(chemSpecies))
+    del myConfig[chemSpeciesTag]     # delete the existing section
+
+    # set up dictionary of chemicals and initial values
+    chemValueDict = {}
+    temperature = 0.0
+    pressure = 0.0
+    for key, value in initValues.items():
+        if (key == "temperature"):
+            temperature = safeFloat(value[valueIndex])
+            continue
+        if (key == "pressure"):
+            pressure = safeFloat(value[valueIndex])
+            continue
+
+        chemValueDict[key] = {
+            "initial value [{}]".format(value[unitIndex]): value[valueIndex]}
+
+    myConfig[chemSpeciesTag] = chemValueDict
+
+    # replace the values of temperature and pressure
+    envConditionsTag = "environmental conditions"
+    envConfig = myConfig[envConditionsTag]
+    envConfig["temperature"]["initial value [K]"] = temperature
+    envConfig["pressure"]["initial value [Pa]"] = pressure
+
+    # save over the former json file
+    with open(myConfigFile, "w") as myConfigFp:
+        json.dump(myConfig, myConfigFp, indent=2)
+
+    # compress the written directory as a zip file
+    shutil.make_archive(destPath, "zip",
+                        root_dir=destDir, base_dir=destZip)
+
+    # move into the created directory
+    shutil.move(destPath + ".zip", destPath)
+
+    return
 
 
 # Main routine begins here.
@@ -236,11 +304,11 @@ def main():
     # get the date-time to retrieve
     dateStr = None
     if ("date" in myArgs):
-      dateStr = myArgs.get("date")
+        dateStr = myArgs.get("date")
 
     timeStr = "00:00"
     if ("time" in myArgs):
-      timeStr = myArgs.get("time")
+        timeStr = myArgs.get("time")
 
     # get the geographical location to retrieve
     lat = None
@@ -252,44 +320,48 @@ def main():
         lon = safeFloat(myArgs.get("longitude"))
 
     retrieveWhen = datetime.datetime.strptime(
-      "{} {}".format(dateStr, timeStr), "%Y%m%d %H:%M")
+        "{} {}".format(dateStr, timeStr), "%Y%m%d %H:%M")
+
+    template = None
+    if ("template" in myArgs):
+        template = myArgs.get("template")
 
     logger.info("Retrieve WACCM conditions at ({} North, {} East)   when {}."
-        .format(lat, lon, retrieveWhen))
+                .format(lat, lon, retrieveWhen))
 
     # Read named variables from WACCM model output.
     varValues = readWACCM(getMusicaDictionary(),
-      lat, lon, retrieveWhen, waccmDir)
+                          lat, lon, retrieveWhen, waccmDir)
     logger.info("Original WACCM varValues = {}".format(varValues))
 
     # Perform any conversions needed, or derive variables.
     varValues = convertWaccm(varValues)
     logger.info("Converted WACCM varValues = {}".format(varValues))
 
-    if (False):
-      # Write CSV file for MusicBox initial conditions.
-      csvName = "{}/{}".format(musicaDir, "initial_conditions.csv")
-      writeInitCSV(varValues, csvName)
+    if (True):
+        # Write CSV file for MusicBox initial conditions.
+        csvName = "{}/{}".format(musicaDir, "initial_conditions.csv")
+        writeInitCSV(varValues, csvName)
 
-    else:
-      # Write JSON file for MusicBox initial conditions.
-      jsonName = "{}/{}".format(musicaDir, "initial_config.json")
-      writeInitJSON(varValues, jsonName)
+    if (True):
+        # Write JSON file for MusicBox initial conditions.
+        jsonName = "{}/{}".format(musicaDir, "initial_config.json")
+        writeInitJSON(varValues, jsonName)
 
-    logger.info("End time: {}".format(datetime.datetime.now()))
-    sys.exit(0)     # no error
-
-
-
-if (__name__ == "__main__"):
-    main()
-
+    if (True and template is not None):
+        logger.info("Insert values into template {}".format(template))
+        insertIntoTemplate(varValues, template, musicaDir)
 
     logger.info("End time: {}".format(datetime.datetime.now()))
     sys.exit(0)     # no error
 
 
-
 if (__name__ == "__main__"):
     main()
 
+    logger.info("End time: {}".format(datetime.datetime.now()))
+    sys.exit(0)     # no error
+
+
+if (__name__ == "__main__"):
+    main()
